// src/generated/pcore/pcore.ts
export interface Version {
    /**
     * @generated from protobuf field: uint32 major = 1
     */
    major: number;
    /**
     * @generated from protobuf field: uint32 minor = 2
     */
    minor: number;
    /**
     * @generated from protobuf field: uint32 patch = 3
     */
    patch: number;
}
/**
 * This is the outer layer.
 *
 * @generated from protobuf message com.preventicus.pcore.Data
 */
export interface Data {
    /**
     * @generated from protobuf field: com.preventicus.pcore.Data.Metadata metadata = 1
     */
    metadata?: Data_Metadata;
    /**
     * @generated from protobuf field: com.preventicus.pcore.Data.CompressedTimestampsContainer compressed_timestamps_container = 2
     */
    compressedTimestampsContainer?: Data_CompressedTimestampsContainer;
    /**
     * @generated from protobuf field: repeated com.preventicus.pcore.Data.Sensor sensors = 3
     */
    sensors: Data_Sensor[];
}
/**
 *
 * Metadata
 *
 * The metadata collects all global data
 *
 * @generated from protobuf message com.preventicus.pcore.Data.Metadata
 */
export interface Data_Metadata {
    /**
     * @generated from protobuf field: sint32 timezone_offset_min = 1
     */
    timezoneOffsetMin: number; // The time zone is set in minutes 
    /**
     * @generated from protobuf field: com.preventicus.pcore.Version pcore_version = 2
     */
    pcoreVersion?: Version; // The version of the used pcore definition
    /**
     * @generated from protobuf field: com.preventicus.pcore.Data.Metadata.Device device = 3
     */
    device?: Data_Metadata_Device; // information about the device
}
/**
 * @generated from protobuf message com.preventicus.pcore.Data.Metadata.Device
 */
export interface Data_Metadata_Device {
    /**
     * @generated from protobuf field: string name = 1
     */
    name: string;
    /**
     * @generated from protobuf field: string manufacturer = 2
     */
    manufacturer: string;
    /**
     * @generated from protobuf field: string id = 3
     */
    id: string;
    /**
     * @generated from protobuf field: com.preventicus.pcore.Version firmware_version = 4
     */
    firmwareVersion?: Version;
}
// 
// Timestamps 

/**
 * @generated from protobuf message com.preventicus.pcore.Data.CompressedTimestampsContainer
 */
export interface Data_CompressedTimestampsContainer {
    /**
     * @generated from protobuf field: uint64 first_unix_timestamp_ms = 1
     */
    firstUnixTimestampMs: number;
    /**
     * @generated from protobuf field: repeated uint32 outer_sections_durations_ms = 2
     */
    outerSectionsDurationsMs: number[];
    /**
     * @generated from protobuf field: repeated uint32 inner_sections_durations_ms = 3
     */
    innerSectionsDurationsMs: number[];
    /**
     * @generated from protobuf field: repeated uint32 sections_sizes = 4
     */
    sectionsSizes: number[];
}
// 
// Sensor

/**
 * @generated from protobuf message com.preventicus.pcore.Data.Sensor
 */
export interface Data_Sensor {
    /**
     * @generated from protobuf oneof: Values
     */
    values: {
        oneofKind: "intValuesContainer";
        /**
         * @generated from protobuf field: com.preventicus.pcore.Data.Sensor.IntValuesContainer int_values_container = 1
         */
        intValuesContainer: Data_Sensor_IntValuesContainer;
    } | {
        oneofKind: "doubleValuesContainer";
        /**
         * @generated from protobuf field: com.preventicus.pcore.Data.Sensor.DoubleValuesContainer double_values_container = 2
         */
        doubleValuesContainer: Data_Sensor_DoubleValuesContainer;
    } | {
        oneofKind: undefined;
    };
    /**
     * @generated from protobuf oneof: Type
     */
    type: {
        oneofKind: "photoplethysmograph";
        /**
         * @generated from protobuf field: com.preventicus.pcore.Data.Sensor.Photoplethysmograph photoplethysmograph = 6
         */
        photoplethysmograph: Data_Sensor_Photoplethysmograph;
    } | {
        oneofKind: "accelerometer";
        /**
         * @generated from protobuf field: com.preventicus.pcore.Data.Sensor.Accelerometer accelerometer = 7
         */
        accelerometer: Data_Sensor_Accelerometer;
    } | {
        oneofKind: "electrocardiogram";
        /**
         * @generated from protobuf field: com.preventicus.pcore.Data.Sensor.Electrocardiogram electrocardiogram = 8
         */
        electrocardiogram: Data_Sensor_Electrocardiogram;
    } | {
        oneofKind: undefined;
    };
}
// 
// Values 

/**
 * @generated from protobuf message com.preventicus.pcore.Data.Sensor.IntValuesContainer
 */
export interface Data_Sensor_IntValuesContainer {
    /**
     * @generated from protobuf field: repeated sint64 values = 1
     */
    values: number[];
}
/**
 * @generated from protobuf message com.preventicus.pcore.Data.Sensor.DoubleValuesContainer
 */
export interface Data_Sensor_DoubleValuesContainer {
    /**
     * @generated from protobuf field: repeated double values = 1
     */
    values: number[];
}
// ...
export interface Data_Sensor_Photoplethysmograph {
    /**
     * @generated from protobuf oneof: Light
     */
    light: {
        oneofKind: "wavelengthNm";
        /**
         * @generated from protobuf field: uint32 wavelength_nm = 1
         */
        wavelengthNm: number;
    } | {
        oneofKind: "color";
        /**
         * @generated from protobuf field: com.preventicus.pcore.Data.Sensor.Photoplethysmograph.Color color = 2
         */
        color: Data_Sensor_Photoplethysmograph_Color;
    } | {
        oneofKind: undefined;
    };
}
// ...
export const Version = new Version$Type();
// @generated message type with reflection information, may provide speed optimized methods
class Data$Type extends MessageType<Data> {
    constructor() {
        super("com.preventicus.pcore.Data", [
            { no: 1, name: "metadata", kind: "message", T: () => Data_Metadata },
            { no: 2, name: "compressed_timestamps_container", kind: "message", T: () => Data_CompressedTimestampsContainer },
            { no: 3, name: "sensors", kind: "message", repeat: 2 /*RepeatType.UNPACKED*/, T: () => Data_Sensor }
        ]);
    }
    create(value?: PartialMessage<Data>): Data {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.sensors = [];
        if (value !== undefined)
            reflectionMergePartial<Data>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: Data): Data {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* com.preventicus.pcore.Data.Metadata metadata */ 1:
                    message.metadata = Data_Metadata.internalBinaryRead(reader, reader.uint32(), options, message.metadata);
                    break;
                case /* com.preventicus.pcore.Data.CompressedTimestampsContainer compressed_timestamps_container */ 2:
                    message.compressedTimestampsContainer = Data_CompressedTimestampsContainer.internalBinaryRead(reader, reader.uint32(), options, message.compressedTimestampsContainer);
                    break;
                case /* repeated com.preventicus.pcore.Data.Sensor sensors */ 3:
                    message.sensors.push(Data_Sensor.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: Data, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* com.preventicus.pcore.Data.Metadata metadata = 1; */
        if (message.metadata)
            Data_Metadata.internalBinaryWrite(message.metadata, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* com.preventicus.pcore.Data.CompressedTimestampsContainer compressed_timestamps_container = 2; */
        if (message.compressedTimestampsContainer)
            Data_CompressedTimestampsContainer.internalBinaryWrite(message.compressedTimestampsContainer, writer.tag(2, WireType.LengthDelimited).fork(), options).join();
        /* repeated com.preventicus.pcore.Data.Sensor sensors = 3; */
        for (let i = 0; i < message.sensors.length; i++)
            Data_Sensor.internalBinaryWrite(message.sensors[i], writer.tag(3, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message com.preventicus.pcore.Data
 */
export const Data = new Data$Type();
// @generated message type with reflection information, may provide speed optimized methods
class Data_Metadata$Type extends MessageType<Data_Metadata> {
    constructor() {
        super("com.preventicus.pcore.Data.Metadata", [
            { no: 1, name: "timezone_offset_min", kind: "scalar", T: 17 /*ScalarType.SINT32*/ },
            { no: 2, name: "pcore_version", kind: "message", T: () => Version },
            { no: 3, name: "device", kind: "message", T: () => Data_Metadata_Device }
        ]);
    }
    create(value?: PartialMessage<Data_Metadata>): Data_Metadata {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.timezoneOffsetMin = 0;
        if (value !== undefined)
            reflectionMergePartial<Data_Metadata>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: Data_Metadata): Data_Metadata {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* sint32 timezone_offset_min */ 1:
                    message.timezoneOffsetMin = reader.sint32();
                    break;
                case /* com.preventicus.pcore.Version pcore_version */ 2:
                    message.pcoreVersion = Version.internalBinaryRead(reader, reader.uint32(), options, message.pcoreVersion);
                    break;
                case /* com.preventicus.pcore.Data.Metadata.Device device */ 3:
                    message.device = Data_Metadata_Device.internalBinaryRead(reader, reader.uint32(), options, message.device);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: Data_Metadata, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* sint32 timezone_offset_min = 1; */
        if (message.timezoneOffsetMin !== 0)
            writer.tag(1, WireType.Varint).sint32(message.timezoneOffsetMin);
        /* com.preventicus.pcore.Version pcore_version = 2; */
        if (message.pcoreVersion)
            Version.internalBinaryWrite(message.pcoreVersion, writer.tag(2, WireType.LengthDelimited).fork(), options).join();
        /* com.preventicus.pcore.Data.Metadata.Device device = 3; */
        if (message.device)
            Data_Metadata_Device.internalBinaryWrite(message.device, writer.tag(3, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message com.preventicus.pcore.Data.Metadata
 */
export const Data_Metadata = new Data_Metadata$Type();
// ...
export const Data_Metadata_Device = new Data_Metadata_Device$Type();
// ...
export const Data_CompressedTimestampsContainer = new Data_CompressedTimestampsContainer$Type();
// ...
export const Data_Sensor = new Data_Sensor$Type();
// ...
export const Data_Sensor_IntValuesContainer = new Data_Sensor_IntValuesContainer$Type();
// ...
export const Data_Sensor_DoubleValuesContainer = new Data_Sensor_DoubleValuesContainer$Type();
// ...
export const Data_Sensor_Photoplethysmograph = new Data_Sensor_Photoplethysmograph$Type();